Objects Starting To Move!!
After setting up the project files and looking at the test movement made for the player (This was to give me a visual to see how i wanted movement to work on an android device).

i got to work on a new movement which takes into acount of player input instead of key presses. The game will be for mobile devices so instead of using UI buttons i just used the touch feature all modern tabloets and phones have and when you swip the screen the player moves.

Also, i added a boat ahead of the player that keeps a distance value between the 2 GameObjects positions. I added the boat as the game is an endless runner which takes place at sea. The boat will drag the player around and the distance is needed so the player does not go in front of the boat.

So instead of the player haviing the movement logic applied the boat does as it gives the illution that the player is controlling the boat which is dragging the player character around.

In my next commit i will be adding obsticles that the player can dodge and slowly working the make the world endless.