Further Movement detections and Out Of Bounds Corrections
In this commit i decided that before working on the environment of the game my movement should be better than just moving left and right.
Although the movement worked fine and it did what it was made for.
I would have eventually needed to make the character move up and down.
Swipping up to jump and down to slide.

Although at the moment i do not have any assets to use with animations to give the effect of jumping and diving.
So i used the players transform and scale to give the effect of these stats of play which will later be exacuted.
I made the Player Object smaller when the player swips down so that the player can fit under object in the way.
Swip up for when the player needs to jump over the objects.

These obsticles are not in place yet however will be soon.

I also decided to clamp the player. This needed to be done as i didnt want the player moving out of bounds which was doable in the new movement system.
As there were no restraints on how far the player can move left or right