Is The Camera And Platforms Moving
So in this commit we have the main character that follows the player. Using a variable that allows the people to edit the camera's focus position through the IDE.

I also made it so that an array of platforms would work together to make an endless path. This is the simple version of how the endless runner will work however the underwater events will be implemented later.

The Next step will be to create a low poly or cel shadded water to make sure the movement looks good with the endless path. However i might just start to implement random obsticles that can stop the player if they are collided with.