I made water
In this commit i made cartoon themed water. I was away for a while on holiday visiting family however, im back now and active on this project. I made the cartoon water with LWPR in unity and soon I'll be making a trail that follows behind the player and concept drawings will be coming soon. They're not great because I can't  draw for shit but it's an insight to what I'm seeing in my mind. So next time you'll see drawings and a trail effect and maybe if the time is right incoming obstacles.