Showing off our shark friend
While I was away dealing with life stuff I decided that I would make my 3D asset for Sea Skater this also will follow with a humanoid Character. I want to keep to a low poly art style with a cartoon appearance. So far my 3D assets are premature and I'm no pro at 3D art however, making my assets is the challenge I want to face and succeed.  All the assets in Sea Skater will be my own and none will be taken from other creators. I'm going to be showing more animations for my Low Poly shark soon as in Sea Skater there are different ways I want it to swim like side to side jumping out of the water or diving down into it. As a side note, none of my own assets will be on GitHub, if I make the shark or human with multiple animations and skins which I plan to do they will be put up on the unity store for a fair price.